requirements.txt
pyYAML==6.0.3
actualpy==0.19.0
numpy==2.4.1
pandas==2.3.3
scikit-learn==1.8.0
rich==14.2.0
beautifulsoup4==4.14.3
google-api-python-client==2.188.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3